+++ src/Http/Controllers/Category/IndexCategoryController.php
@@ -8,12 +8,7 @@ class IndexCategoryController extends Controller
 {
     public function index($slug = 'ultimas-noticias')
     {
-        $data = [
-            'slug' => $slug,
-            'editoria' => 'ultimas-noticias',
-        ];
-
-        return view('category.index', $data);
+        return view('category.index', [ 'slug' => $slug ]);
     }
 
     public function listByCategory($slug)
